+++ CultWeb/src/main/java/ch/unige/idsi/cultweb/model/Museum.java
@@ -4,13 +4,12 @@ import java.util.List;
 
 public class Museum extends Place {
 	
-	public Museum(long id, String name, String contact, String town, String address, String url, long latitude, long longitude) {
+	public Museum(int id, String name, String contact, String town, String address, String url, double latitude, double longitude) {
 		super(id, name, contact, town, address, url, Infrastructure.MUSEUM, latitude, longitude);
 	}
 	
-	public Museum(long id, String name, String contact, String town, String address, String url, long latitude, long longitude, List<Recommendation> recommendations) {
+	public Museum(int id, String name, String contact, String town, String address, String url, double latitude, double longitude, List<Recommendation> recommendations) {
 		super(id, name, contact, town, address, url, Infrastructure.MUSEUM, latitude, longitude, recommendations);
 	}
-	
 	
 }
